split MovingPointIdentification
